requirements-doc.txt
-r requirements.txt

# document building requirements
Jinja2
numpydoc>=0.6
sphinx>=1.5.4
sphinx_rtd_theme
